backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
openai==1.61.1

